Fix possible division casting issue (#8681)

+++ radio/src/telemetry/ghost.cpp
@@ -271,8 +271,8 @@ void processGhostTelemetryFrame()
         bluetooth.write(telemetryRxBuffer, telemetryRxBufferCount);
       }
 #endif
-      processGhostTelemetryValue(GHOST_ID_GPS_LAT, getTelemetryValue_s32le(3) / 10);  
-      processGhostTelemetryValue(GHOST_ID_GPS_LONG, getTelemetryValue_s32le(7) / 10);
+      processGhostTelemetryValue(GHOST_ID_GPS_LAT, ((int32_t)getTelemetryValue_s32le(3)) / 10);
+      processGhostTelemetryValue(GHOST_ID_GPS_LONG, ((int32_t)getTelemetryValue_s32le(7)) / 10);
       processGhostTelemetryValue(GHOST_ID_GPS_ALT, (int16_t) getTelemetryValue_u16le(11));  
       break; 
     }
